# backend/docs/openapi-3.yaml
components:
  schemas:
...
    api.ProxyPoolMembershipResponse:
      properties:
        added:
          type: boolean
        message:
          type: string
        poolId:
          type: string
        proxyId:
          type: string
        removed:
          type: boolean
      type: object
    api.ProxyPoolRequest:
      properties:
        description:
          type: string
        healthCheckEnabled:
          type: boolean
        healthCheckIntervalSeconds:
          type: integer
        isEnabled:
          type: boolean
        maxRetries:
          type: integer
        name:
          type: string
        poolStrategy:
          type: string
        timeoutSeconds:
          type: integer
      required:
      - name
      type: object
...
    models.Proxy:
      properties:
        address:
          description: Full proxy address (e.g., 'http://user:pass@host:port')
          type: string
        city:
          $ref: '#/components/schemas/sql.NullString'
        countryCode:
          $ref: '#/components/schemas/sql.NullString'
        createdAt:
          type: string
        description:
          $ref: '#/components/schemas/sql.NullString'
        failureCount:
          $ref: '#/components/schemas/sql.NullInt32'
        host:
          $ref: '#/components/schemas/sql.NullString'
        id:
          type: string
        inputPassword:
          $ref: '#/components/schemas/sql.NullString'
        inputUsername:
          $ref: '#/components/schemas/sql.NullString'
        isEnabled:
          type: boolean
        isHealthy:
          type: boolean
        lastCheckedAt:
          $ref: '#/components/schemas/sql.NullTime'
        lastError:
          $ref: '#/components/schemas/sql.NullString'
        lastStatus:
          $ref: '#/components/schemas/sql.NullString'
        lastTested:
          $ref: '#/components/schemas/sql.NullTime'
        latencyMs:
          $ref: '#/components/schemas/sql.NullInt32'
        name:
          type: string
        notes:
          $ref: '#/components/schemas/sql.NullString'
        port:
          $ref: '#/components/schemas/sql.NullInt32'
        protocol:
          $ref: '#/components/schemas/models.ProxyProtocolEnum'
        provider:
          $ref: '#/components/schemas/sql.NullString'
        status:
          $ref: '#/components/schemas/models.ProxyStatusEnum'
        successCount:
          $ref: '#/components/schemas/sql.NullInt32'
        updatedAt:
          type: string
        username:
          $ref: '#/components/schemas/sql.NullString'
      required:
      - address
      - name
      type: object
    models.ProxyPool:
      properties:
        createdAt:
          type: string
        description:
          $ref: '#/components/schemas/sql.NullString'
        healthCheckEnabled:
          type: boolean
        healthCheckIntervalSeconds:
          type: integer
        id:
          type: string
        isEnabled:
          type: boolean
        maxRetries:
          type: integer
        name:
          type: string
        poolStrategy:
          $ref: '#/components/schemas/sql.NullString'
        proxies:
          description: Computed fields (not stored in DB)
          items:
            $ref: '#/components/schemas/models.Proxy'
          type: array
          uniqueItems: false
        timeoutSeconds:
          type: integer
        updatedAt:
          type: string
      required:
      - name
      type: object
    models.ProxyPoolMembership:
      properties:
        addedAt:
          type: string
        isActive:
          type: boolean
        poolId:
          type: string
        proxyId:
          type: string
        weight:
          type: integer
      required:
      - poolId
      - proxyId
      type: object
    models.ProxyProtocolEnum:
      type: string
      x-enum-comments:
        ProxyProtocolHTTP: '@enum http'
        ProxyProtocolHTTPS: '@enum https'
        ProxyProtocolSOCKS4: '@enum socks4'
        ProxyProtocolSOCKS5: '@enum socks5'
      x-enum-varnames:
      - ProxyProtocolHTTP
      - ProxyProtocolHTTPS
      - ProxyProtocolSOCKS5
      - ProxyProtocolSOCKS4
    models.ProxyStatusEnum:
      description: Frontend-expected fields
      type: string
      x-enum-varnames:
      - ProxyStatusActive
      - ProxyStatusDisabled
      - ProxyStatusTesting
      - ProxyStatusFailed
...
    sql.NullInt32:
      description: Port number
      properties:
        int32:
          type: integer
        valid:
          description: Valid is true if Int32 is not NULL
          type: boolean
      type: object
    sql.NullString:
      description: Fields for input/logic, not direct DB columns if already covered
        by Address or PasswordHash
      properties:
        string:
          type: string
        valid:
          description: Valid is true if String is not NULL
          type: boolean
      type: object
    sql.NullTime:
      description: Timestamp of last health check
      properties:
        time:
          type: string
        valid:
          description: Valid is true if Time is not NULL
          type: boolean
      type: object
...
  securitySchemes:
    basic:
      scheme: basic
      type: http
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for domain generation, validation, and management
    operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Domain Flow API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /proxy-pools:
    get:
      description: Retrieve all proxy pools with their associated proxies
      operationId: listProxyPools
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.ProxyPool'
                type: array
          description: List of proxy pools
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list proxy pools
      summary: List proxy pools
      tags:
      - proxy-pools
    post:
      description: Create a new proxy pool with configuration settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.ProxyPoolRequest'
        description: Proxy pool creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProxyPool'
          description: Created proxy pool
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request payload
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create pool
      summary: Create proxy pool
      tags:
      - proxy-pools
  /proxy-pools/{poolId}:
    delete:
      description: Delete a proxy pool
      parameters:
      - description: Proxy pool ID
        in: path
        name: poolId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
          description: Deletion confirmation
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to delete pool
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to delete pool
      summary: Delete proxy pool
      tags:
      - proxy-pools
    put:
      description: Update an existing proxy pool configuration
      parameters:
      - description: Proxy pool ID
        in: path
        name: poolId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.ProxyPoolRequest'
        description: Proxy pool update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProxyPool'
          description: Updated proxy pool
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid ID or request payload
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Pool not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to update pool
      summary: Update proxy pool
      tags:
      - proxy-pools
  /proxy-pools/{poolId}/proxies:
    post:
      description: Assign a proxy to a proxy pool with optional weight
      parameters:
      - description: Proxy pool ID
        in: path
        name: poolId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                proxyId:
                  type: string
                weight:
                  type: integer
              type: object
        description: Proxy assignment request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ProxyPoolMembership'
          description: Created membership
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid pool ID, payload, or proxy ID
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to add proxy
      summary: Add proxy to pool
      tags:
      - proxy-pools
  /proxy-pools/{poolId}/proxies/{proxyId}:
    delete:
      description: Remove a proxy from a specific proxy pool
      parameters:
      - description: Pool ID (UUID)
        in: path
        name: poolId
        required: true
        schema:
          type: string
      - description: Proxy ID (UUID)
        in: path
        name: proxyId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ProxyPoolMembershipResponse'
          description: Proxy removed from pool successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Pool or proxy not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Remove proxy from pool
      tags:
      - proxy-pools
...
servers:
- url: /api/v2